test: Implemented tests for create, update, delete funcionality

+++ src/Modules/Tours/Explorer.Tours.API/Dtos/TourReviewDto.cs
@@ -5,13 +5,13 @@ namespace Explorer.Tours.API.Dtos
 {
     public class TourReviewDto
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public int Rating { get; set; }
         public string Comment { get; set; }
         public DateTime VisitDate { get; set; }
         public DateTime ReviewDate { get; set; }
         public string? Image { get; set; }
-        public int TourId { get; set; }
-        public int TouristId { get; set; }
+        public long TourId { get; set; }
+        public long TouristId { get; set; }
     }
 }
